Add _teig
Rename tsvd to _tsvd and add higher level tsvd

Move tsvd to svd.jl and common stuff to common.jl

Add AtA type for computing the SVD from the eigenvalues of the AtA matrix
